Show messages of selected dummy data
