+++ src/Components/Pages/Chat/ChatRoomList.js
@@ -2,21 +2,7 @@ import React, { useState } from 'react';
 import { Paper, IconButton, InputBase, Icon } from '@material-ui/core';
 import ChatRoomBox from './ChatRoomBox';
 
-const dummyMessages = [
-  { name: 'Dummy Name', message: 'This is a message', time: '11:50', unread: '2' },
-  {
-    name: 'My Name',
-    message: 'This is a message more than 50 letters. I am long very long',
-    time: '18:10',
-    unread: '1',
-  },
-  { name: 'Your Name', message: 'This is a message33', time: '04/02', unread: null },
-  { name: 'Their Name', message: 'This is a message44', time: '04/01', unread: null },
-  { name: 'Our Name', message: 'This is a message55', time: '03/31', unread: null },
-  { name: 'No Name', message: 'This is a message66', time: '03/15', unread: null },
-];
-
-export default function ChatRoomList() {
+export default function ChatRoomList({ data, setSelectedRoom }) {
   const [searchInput, setSearchInput] = useState('');
 
   return (
@@ -41,8 +27,8 @@ export default function ChatRoomList() {
         </Paper>
       </div>
 
-      {dummyMessages.map((msg) => (
-        <ChatRoomBox {...msg} />
+      {data.map((msg, index) => (
+        <ChatRoomBox {...msg} index={index} setSelectedRoom={(idx) => setSelectedRoom(idx)} />
       ))}
     </div>
   );
